Move code config to shared code
